+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Next Release
+
++ [**ENHANCEMENT**]**;** Magicxx: Throw filesystem_error from identify_files(directory) if the underlying std::filesystem OS API fails.
+
 + [**ENHANCEMENT**]**;** Magicxx: Add filesystem_error.
 
 + [**QUALITY**]**;** Magicxx: Refactor identify_file(s).